update: move pkg folder to root to match the standard
