+++ go.mod
@@ -1,22 +1,22 @@
 module github.com/sj0n/heepno
 
-go 1.24.0
+go 1.24.5
 
 require (
 	github.com/AssemblyAI/assemblyai-go-sdk v1.10.0
-	github.com/deepgram/deepgram-go-sdk v1.8.2
-	github.com/sashabaranov/go-openai v1.37.0
+	github.com/deepgram/deepgram-go-sdk v1.9.0
+	github.com/sashabaranov/go-openai v1.40.5
 	github.com/spf13/cobra v1.9.1
 )
 
 require (
 	github.com/cenkalti/backoff v2.2.1+incompatible // indirect
-	github.com/coder/websocket v1.8.12 // indirect
+	github.com/coder/websocket v1.8.13 // indirect
 	github.com/dvonthenen/websocket v1.5.1-dyv.2 // indirect
-	github.com/go-logr/logr v1.4.2 // indirect
+	github.com/go-logr/logr v1.4.3 // indirect
 	github.com/google/go-querystring v1.1.0 // indirect
 	github.com/gorilla/schema v1.4.1 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/pflag v1.0.6 // indirect
+	github.com/spf13/pflag v1.0.7 // indirect
 	k8s.io/klog/v2 v2.130.1 // indirect
 )
